requirements.txt.txt
streamlit
pandas
gspread
google-auth
openpyxl 
reportlab